refactor(hostsvc): split handlers.go into per-tier files (#350)
Mechanically split the 975-LOC internal/plugin/hostsvc/handlers.go into
five focused same-package files, with zero behavior, signature, or receiver
changes — every declaration moves verbatim:

- helpers.go            shared helpers (resolveInstance, latestStepNumber, writeAuditEvent)
- policy_scope.go       pure policy-inspection (scopeProbe, policyGrantsInstance, policyIDsForInstance)
- handlers_feedback.go  security-critical audit/feedback path (WriteAuditStep, writeFeedbackResponseStep)
- handlers_tier2.go     Tier-2 data-access RPCs (RunHistoryRead, UserDirectoryRead, hasTier2Capability)
- handlers_tier1.go     Tier-1 simple RPCs + enum mappers

Implemented via agentic development pipeline.

Closes #350
